Conflicts:
	SConstruct

+++ lib/include/fx/easing/BackEase.h
@@ -17,7 +17,7 @@ namespace stm32plus {
     class BackEase : public EasingBase {
 
       private:
-        double _overshoot;
+        float _overshoot;
 
       public:
         /// constructor - sets a default value for the overshoot
@@ -25,21 +25,21 @@ namespace stm32plus {
 
         /// Starts the motion by backtracking, then reversing
         /// direction and moving toward the target
-        virtual double easeIn(double time) const override;
+        virtual float easeIn(float time) const override;
 
         /// Starts the motion by moving towards the target, overshooting
         /// it slightly, and then reversing direction back toward the target
-        virtual double easeOut(double time) const override;
+        virtual float easeOut(float time) const override;
 
         /// Combines the motion of the easeIn and easeOut methods to
         /// start the motion by backtracking, then reversing direction
         /// and moving toward target, overshooting target slightly,
         /// reversing direction again, and then moving back toward the target
-        virtual double easeInOut(double time) const override;
+        virtual float easeInOut(float time) const override;
 
         /// set the overshoot value. The higher the value the
         /// greater the overshoot.
-        void setOvershoot(double overshoot);
+        void setOvershoot(float overshoot);
     };
   }
 }
